requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.2.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
pyotp==2.9.0
bcrypt==4.0.1
Flask-Mail==0.9.1
Click==8.1.7
python-dotenv==1.0.0
qrcode==7.4.2
tabulate==0.9.0
pillow==10.0.0
pytest==7.4.0
psutil==5.9.5 